Move Divisualisation spacetime renderer to spacetime.py
Keep divisualisation.py as a thin back-compat shim that re-exports it.
Fold in the OpenGL render-context fix from fix/render-opengl-context and
drop a dead np.zeros_like line.
